Fix "optimization"
This optimization was originally just a typo, and it did not appear to affect the output at all. Upon further investigation, it does -- significantly, in fact, when using percentile values other than 50 (p=0.5).

This switches to an optimization that only assumes that `di=0` is rare, impossible, or unimportant. This is plausible for `Linear()` because it always divides-by-zero in that case (easily avoided in the scalar code with a branch). This saves 2 calls to `Select()` for each sample.